Split psrc and errset
Split out psrc and errset
