+++ internal/j5s/protobuild/linker_test.go
@@ -4,36 +4,32 @@ import (
 	"testing"
 
 	"github.com/bufbuild/protocompile/linker"
+	"github.com/pentops/j5/internal/j5s/protobuild/psrc"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/descriptorpb"
 )
 
 func TestLinkerDeps(t *testing.T) {
 	ctx := t.Context()
 
-	tf := &testDeps{
-		externalDeps: map[string]*descriptorpb.FileDescriptorProto{
-			"test/v1/foo.proto": {
-				Name:    proto.String("test/v1/foo.proto"),
-				Syntax:  proto.String("proto3"),
-				Package: proto.String("test.v1"),
-				Dependency: []string{
-					"google/protobuf/timestamp.proto",
-				},
+	tf := testDeps{
+		"test/v1/foo.proto": {
+			Name:    proto.String("test/v1/foo.proto"),
+			Syntax:  proto.String("proto3"),
+			Package: proto.String("test.v1"),
+			Dependency: []string{
+				"google/protobuf/timestamp.proto",
 			},
 		},
 	}
 
-	deps, err := newDependencyResolver(tf)
+	deps, err := psrc.ChainResolver(tf)
 	if err != nil {
 		t.Fatalf("FATAL: Unexpected error: %s", err.Error())
 	}
 
-	resolver := newResolverCache(newBuiltinResolver(), deps)
-
 	filenames := []string{"test/v1/foo.proto"}
 
-	cc := newLinker(resolver, &linker.Symbols{})
+	cc := newLinker(deps, &linker.Symbols{})
 	files, err := cc.resolveAll(ctx, filenames)
 	if err != nil {
 		t.Fatalf("FATAL: Unexpected error: %s", err.Error())
